requirements.txt
bs4
requests
pytest
pylint
python-dotenv
psycopg2-binary
pandas
boto3
streamlit
pandas
altair